requirements.txt

fastapi==0.111.0
langchain==0.2.5
langchain-community==0.2.5
langchain-core==0.2.7
langchain-google-genai==1.0.6
langchain-qdrant==0.1.1
langchain-text-splitters==0.2.1
PyPDF2==3.0.1
python-docx==1.1.2
python-dotenv==1.0.1
streamlit==1.35.0
tiktoken==0.7.0
